# services/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Ocserv User management Api server.",
        "title": "Ocserv User management Example Api",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/customers/setup/cisco": {
            "post": {
                "description": "Create Cisco Secure Client certificate import and connection creation URIs using ocserv username/password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customers"
                ],
                "summary": "Create customer Cisco Secure Client setup links",
                "parameters": [
                    {
                        "description": "customer username and password (same ocserv account).",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/customer.SummaryData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/customer.CiscoSetupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/middlewares.TooManyRequests"
                        }
                    }
                }
            }
        },
        "/customers/setup/cisco/certificate/{token}": {
            "get": {
                "description": "Download customer's PKCS#12 certificate using a short-lived Cisco Secure Client setup token",
                "produces": [
                    "application/x-pkcs12"
                ],
                "tags": [
                    "Customers"
                ],
                "summary": "Download customer Cisco Secure Client setup certificate",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cisco Secure Client setup certificate token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user.p12",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/middlewares.TooManyRequests"
                        }
                    }
                }
            }
        },
        "/customers/setup/cisco/launch/certificate/{token}": {
            "get": {
                "description": "Redirect a short-lived signed HTTPS setup link to the Cisco Secure Client certificate import URI",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Customers"
                ],
                "summary": "Launch Cisco certificate import",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cisco Secure Client setup token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/middlewares.TooManyRequests"
                        }
                    }
                }
            }
        },
        "/customers/setup/cisco/launch/connection/{token}": {
            "get": {
                "description": "Redirect a short-lived signed HTTPS setup link to the Cisco Secure Client connection creation URI",
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Customers"
                ],
                "summary": "Launch Cisco connection creation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cisco Secure Client setup token",
                        "name": "token",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/middlewares.TooManyRequests"
                        }
                    }
                }
            }
        },
...
        "/gateway/users/{username}/cisco-setup": {
            "post": {
                "description": "Creates Cisco Secure Client certificate import and connection creation URIs for an existing ocserv user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Gateway"
                ],
                "summary": "Gateway Cisco Secure Client setup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer GATEWAY_API_TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Ocserv username",
                        "name": "username",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gateway.CiscoSetupResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Unauthorized"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "customer.CiscoSetupResponse": {
            "type": "object",
            "required": [
                "certificate_import_uri",
                "certificate_import_url",
                "certificate_password",
                "connection_create_uri",
                "connection_create_url",
                "connection_name",
                "expires_at",
                "server_address",
                "server_port"
            ],
            "properties": {
                "certificate_import_uri": {
                    "type": "string"
                },
                "certificate_import_url": {
                    "type": "string"
                },
                "certificate_password": {
                    "type": "string"
                },
                "connection_create_uri": {
                    "type": "string"
                },
                "connection_create_url": {
                    "type": "string"
                },
                "connection_name": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "server_address": {
                    "type": "string"
                },
                "server_port": {
                    "type": "integer"
                }
            }
        },
...
        "customer.SummaryData": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 2
                },
                "username": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 2
                }
            }
        },
...
        "gateway.CiscoSetupResponse": {
            "type": "object",
            "properties": {
                "certificate_import_uri": {
                    "type": "string"
                },
                "certificate_import_url": {
                    "type": "string"
                },
                "certificate_password": {
                    "type": "string"
                },
                "connection_create_uri": {
                    "type": "string"
                },
                "connection_create_url": {
                    "type": "string"
                },
                "connection_name": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "server_address": {
                    "type": "string"
                },
                "server_port": {
                    "type": "integer"
                }
            }
        },
...
        "middlewares.TooManyRequests": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "middlewares.Unauthorized": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "request.ErrorResponse": {
            "type": "object",
            "required": [
                "error",
                "message"
            ],
            "properties": {
                "error": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    }
}