# services/api/internal/services/gateway/types.go
package gateway

import "time"

type CreateUserData struct {
	Username       string  `json:"username" validate:"required,min=2,max=64"`
	Password       string  `json:"password" validate:"required,min=4,max=64"`
	Group          string  `json:"group" validate:"required"`
	Unlimited      bool    `json:"unlimited" validate:"omitempty"`
	ExpireAt       *string `json:"expire_at" validate:"omitempty" example:"2026-12-31"`
	TrafficType    string  `json:"traffic_type" validate:"required,oneof=Free MonthlyTransmit MonthlyReceive MonthlyRxTx TotallyTransmit TotallyReceive TotallyRxTx"`
	TrafficLimitGB int     `json:"traffic_limit_gb" validate:"omitempty,min=0,max=100000"`
	Description    string  `json:"description" validate:"omitempty,max=1024"`
}

type CreateUserResponse struct {
	RemoteUserID   string  `json:"remote_user_id"`
	Username       string  `json:"username"`
	Password       string  `json:"password"`
	Group          string  `json:"group"`
	Unlimited      bool    `json:"unlimited"`
	ExpireAt       *string `json:"expire_at,omitempty"`
	TrafficType    string  `json:"traffic_type"`
	TrafficLimitGB int     `json:"traffic_limit_gb"`
}

type UserStatusResponse struct {
	RemoteUserID          string  `json:"remote_user_id"`
	Username              string  `json:"username"`
	Group                 string  `json:"group"`
	Active                bool    `json:"active"`
	Locked                bool    `json:"locked"`
	Deactivated           bool    `json:"deactivated"`
	Unlimited             bool    `json:"unlimited"`
	ExpireAt              *string `json:"expire_at,omitempty"`
	DeactivatedAt         *string `json:"deactivated_at,omitempty"`
	TrafficType           string  `json:"traffic_type"`
	TrafficLimitGB        float64 `json:"traffic_limit_gb"`
	TrafficConsumedGB     float64 `json:"traffic_consumed_gb"`
	TrafficRemainingGB    float64 `json:"traffic_remaining_gb"`
	RxGB                  float64 `json:"rx_gb"`
	TxGB                  float64 `json:"tx_gb"`
	TrafficLimitBytes     int64   `json:"traffic_limit_bytes"`
	TrafficConsumedBytes  int64   `json:"traffic_consumed_bytes"`
	TrafficRemainingBytes int64   `json:"traffic_remaining_bytes"`
	RxBytes               int64   `json:"rx_bytes"`
	TxBytes               int64   `json:"tx_bytes"`
}

type UpdateUserSubscriptionData struct {
	TrafficLimitGB    *int    `json:"traffic_limit_gb" validate:"omitempty,min=1,max=100000"`
	ExpireAt          *string `json:"expire_at" validate:"omitempty" example:"2026-12-31"`
	Unlimited         bool    `json:"unlimited" validate:"omitempty"`
	ResetTrafficUsage bool    `json:"reset_traffic_usage" validate:"omitempty"`
	Activate          bool    `json:"activate" validate:"omitempty"`
	Group             *string `json:"group" validate:"omitempty,min=1,max=16"`
}

type DeleteUserResponse struct {
	RemoteUserID string `json:"remote_user_id"`
	Username     string `json:"username"`
	Deleted      bool   `json:"deleted"`
}

type CiscoSetupResponse struct {
	CertificateImportURI string    `json:"certificate_import_uri"`
	CertificateImportURL string    `json:"certificate_import_url"`
	ConnectionCreateURI  string    `json:"connection_create_uri"`
	ConnectionCreateURL  string    `json:"connection_create_url"`
	CertificatePassword  string    `json:"certificate_password"`
	ConnectionName       string    `json:"connection_name"`
	ServerAddress        string    `json:"server_address"`
	ServerPort           int       `json:"server_port"`
	ExpiresAt            time.Time `json:"expires_at"`
}